requirements.txt
flask==3.1.0
python-dotenv==1.0.1
browser-use==0.1.40
langchain==0.3.21
langchain-anthropic==0.3.3
anthropic==0.49.0
playwright==1.51.0
pymysql==1.1.0
bcrypt==4.0.1
Werkzeug==3.1.3
flask-mail==0.9.1
boto3==1.28.65
pytest==7.4.2